feat: implement image upload functionality and loading state in card generator
